requirements.txt
aiogram==3.15.0
httpx[socks]==0.28.1
anthropic==0.43.0
aiosqlite==0.20.0
APScheduler==3.10.4
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml==5.3.0
